Making FlashIAP object SingletonPtr
Any object left lingering in the source code with static initializer will get pulled in even if not being used.
That's why FlashIAP object should be wrapped in SingeltonPtr envelope.
This will save about 660 B of ROM and 40 B of RAM for applications who do have mbed-clount-client library but are not using it.