Fix pointer argument passed to ARM_UC_cryptoHMACSHA256
At this call to ARM_UC_cryptoHMACSHA256 a pointer to uint32_t is
passed as the last argument. Though this function accepts a
pointer to size_t. On 64 bit machines size_t is 64 bit. This call
causes ARM_UC_cryptoHMACSHA256 to assign to a 64 bit space instead
of actual 32 bit, causing memory corruption and eventual crash.