Bug fix:  Allow empty multi instance resources
The client fails to register currently if a multi instance resource is created without any instances.  This is allowed by the spec, in fact, it makes sense.  The client may want to create a multi instance resource with no initial instances so that the server can PUT the resource instances later.  (For example, our use case is just a simple place for our web team to store information specific to the IoT node in STRING format)