Delete the manifest in on_reboot or on_reboot_in_ms functions.
If the manifest is deleted before those functions, there is a corner case where the FOTA resource state update to FOTA_SOURCE_STATE_REBOOTING isn't completed (i.e. the device doesn't receive and ACK).
In such case, the resume will happen after re-registration, but since the manifest is deleted the resume will fail and the fota will be stuck.
Deleting the manifest only after the fota resource state was being updated, prevents this corner case.

In case the reboot isn't needed, the manifest is deleted after the installation is completed.